# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Simulator API
  version: 0.1.0
  description: |
    REST API for the general-purpose distributed simulation platform.
    Endpoints cover the model registry, simulation lifecycle, snapshots, and
    live state/metrics.
paths:
...
  /api/v1/models/{id}:
    get:
      summary: Get a model by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Model metadata
        '404':
          description: Not found
  /api/v1/simulations:
    get:
      summary: List simulations
      responses:
        '200':
          description: Simulation list
    post:
      summary: Create a simulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model_id, seed]
              properties:
                id: { type: string }
                model_id: { type: string }
                model_version: { type: string }
                seed: { type: integer }
                mode: { type: string, enum: [tick, event] }
                max_ticks: { type: integer }
                max_time: { type: number }
                workers: { type: integer }
                config: { type: object }
      responses:
        '201':
          description: Created simulation record
        '400':
          description: Invalid request
  /api/v1/simulations/{id}:
    get:
      summary: Get a simulation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Simulation record
        '404':
          description: Not found
    delete:
      summary: Delete a simulation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
...